+++ src/server.jl
@@ -489,6 +489,7 @@ function _extract_relevant_options(file_frontmatter::Dict, options::Dict)
             params = params_default,
             cache = cache_default,
             env = Dict{String,Any}(),
+            cwd = nothing,
         )
     else
         format = get(D, options, "format")
@@ -518,6 +519,8 @@ function _extract_relevant_options(file_frontmatter::Dict, options::Dict)
         cli_params = get(options, "params", Dict())
         params_merged = _recursive_merge(params_default, params, cli_params)
 
+        cwd = get(options, "cwd", nothing)
+
         return _options_template(;
             fig_width,
             fig_height,
@@ -531,6 +534,7 @@ function _extract_relevant_options(file_frontmatter::Dict, options::Dict)
             params = params_merged,
             cache,
             env,
+            cwd,
         )
     end
 end
@@ -548,6 +552,7 @@ function _options_template(;
     params,
     cache,
     env,
+    cwd,
 )
     D = Dict{String,Any}
     return D(
@@ -567,6 +572,7 @@ function _options_template(;
         ),
         "params" => D(params),
         "env" => env,
+        "cwd" => cwd,
     )
 end